Refactor Person grid cell in view as re-usable component
