+++ src/features/views/components/ViewDataTable/columnTypes/LocalPersonColumnType.tsx
@@ -1,32 +1,13 @@
-import { FormattedMessage } from 'react-intl';
-import { makeStyles } from '@mui/styles';
-import { useRouter } from 'next/router';
-import {
-  Avatar,
-  Box,
-  Button,
-  InputBase,
-  List,
-  ListItem,
-  ListSubheader,
-  Paper,
-  Popper,
-  Theme,
-  Typography,
-  useAutocomplete,
-} from '@mui/material';
-
 import { IColumnType } from '.';
-import useAccessLevel from 'features/views/hooks/useAccessLevel';
-import { usePersonSelect } from 'zui/ZUIPersonSelect';
 import useViewDataModel from 'features/views/hooks/useViewDataModel';
 import ViewDataModel from 'features/views/models/ViewDataModel';
-import ZUIAvatar from 'zui/ZUIAvatar';
-import { Close, Person } from '@mui/icons-material';
+import ZUIPersonGridCell from 'zui/ZUIPersonGridCell';
+import ZUIPersonGridEditCell from 'zui/ZUIPersonGridEditCell';
+
 import { COLUMN_TYPE, LocalPersonViewColumn } from '../../types';
-import { FC, HTMLAttributes, useState } from 'react';
 import { GridColDef, useGridApiContext } from '@mui/x-data-grid-pro';
 import { ZetkinPerson, ZetkinViewRow } from 'utils/types/zetkin';
+import { FC } from 'react';
 
 type LocalPersonViewCell = null | ZetkinPerson;
 
@@ -46,10 +27,10 @@ export default class LocalPersonColumnType
       headerAlign: 'center',
 
       renderCell: (params) => {
-        return <ReadCell cell={params.value} />;
+        return <ZUIPersonGridCell cell={params.value} />;
       },
       renderEditCell: (params) => {
-        return <Cell cell={params.value} column={col} row={params.row} />;
+        <EditCell cell={params.value} col={col} row={params.row} />;
       },
       sortComparator: (
         val0: LocalPersonViewCell,
@@ -84,256 +65,33 @@ export default class LocalPersonColumnType
   }
 }
 
-const useStyles = makeStyles<Theme, { isRestrictedMode: boolean }>({
-  popper: {
-    display: 'flex',
-    flexDirection: 'column',
-    height: (props) => (props.isRestrictedMode ? 'auto' : 400),
-    width: 300,
-  },
-  searchingList: {
-    height: 'calc(100% - 40px)',
-    minWidth: '250px',
-    overflowY: 'scroll',
-    paddingLeft: '20px',
-    position: 'absolute',
-    width: '100%',
-  },
-});
-
-const ReadCell: FC<{
-  cell: LocalPersonViewCell | undefined;
-}> = ({ cell }) => {
-  const query = useRouter().query;
-  const orgId = parseInt(query.orgId as string);
-
-  return (
-    <Box>
-      {cell ? (
-        <ZUIAvatar orgId={orgId} personId={cell.id} />
-      ) : (
-        <Avatar>
-          <Person />
-        </Avatar>
-      )}
-    </Box>
-  );
-};
-
-const Cell: FC<{
-  cell: LocalPersonViewCell | undefined;
-  column: LocalPersonViewColumn;
+const EditCell: FC<{
+  cell: ZetkinPerson | null;
+  col: LocalPersonViewColumn;
   row: ZetkinViewRow;
-}> = ({ cell, column, row }) => {
-  const query = useRouter().query;
-  const [isRestrictedMode] = useAccessLevel();
-  const styles = useStyles({ isRestrictedMode });
-  const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
-  const [searching, setSearching] = useState(false);
-
+}> = (cell, col, row) => {
+  const api = useGridApiContext();
   const model = useViewDataModel();
 
-  const orgId = parseInt(query.orgId as string);
-
-  const api = useGridApiContext();
-
-  const updateCellValue = (person: ZetkinPerson | null) => {
-    api.current.stopCellEditMode({ field: 'col_' + column.id, id: row.id });
-    model.setCellValue(row.id, column.id, person?.id ?? null);
-    setSearching(false);
+  const getPeopleInView = (people: ZetkinPerson[]) => {
+    return usePeopleInView(model, people);
   };
 
-  const personSelect = usePersonSelect({
-    initialValue: cell ? cell.first_name + ' ' + cell.last_name : '',
-    onChange: (person) => {
-      updateCellValue(person);
-      setAnchorEl(null);
-      setSearching(false);
-    },
-    selectedPerson: null,
-  });
-
-  const autoComplete = useAutocomplete({
-    ...personSelect.autoCompleteProps,
-    disabled: isRestrictedMode,
-  });
-
-  const options = autoComplete.groupedOptions as ZetkinPerson[];
-  const peopleInView = usePeopleInView(model, options);
-  const showPeopleInView = searching || !cell;
+  const updateCellValue = (person: ZetkinPerson | null) => {
+    api.current.stopCellEditMode({
+      field: 'col_' + col.id,
+      id: row.id,
+    });
+    model.setCellValue(row.id, col.id, person?.id ?? null);
+  };
 
   return (
-    <Box
-      onMouseEnter={(ev) => {
-        setAnchorEl(ev.currentTarget);
-      }}
-      onMouseLeave={() => {
-        setAnchorEl(null);
-      }}
-    >
-      <Box ref={(div: HTMLDivElement) => setAnchorEl(div)}>
-        <InputBase
-          /* eslint-disable-next-line jsx-a11y/no-autofocus */
-          autoFocus
-          fullWidth
-          inputProps={autoComplete.getInputProps()}
-          onChange={() => setSearching(true)}
-        ></InputBase>
-      </Box>
-      <Popper
-        anchorEl={anchorEl}
-        open={!!anchorEl}
-        popperOptions={{
-          placement: 'bottom',
-        }}
-      >
-        <Paper
-          className={styles.popper}
-          elevation={2}
-          onClick={(ev) => {
-            ev.stopPropagation();
-            anchorEl?.focus();
-          }}
-        >
-          {isRestrictedMode && (
-            <Box
-              alignItems="center"
-              display="flex"
-              flexDirection="column"
-              gap={1}
-              justifyContent="center"
-              width="100%"
-            >
-              {!!cell && <SelectedPerson orgId={orgId} person={cell} />}
-              <Typography fontStyle="italic" variant="caption">
-                <FormattedMessage id="misc.views.cells.localPerson.restrictedMode" />
-              </Typography>
-            </Box>
-          )}
-          {!isRestrictedMode && (
-            <Box display="flex" flexDirection="column" height="100%">
-              <Box
-                sx={{
-                  alignItems: 'center',
-                  display: 'flex',
-                  height: 'calc(100% - 10px)',
-                  justifyContent: 'center',
-                  minWidth: '290px',
-                  width: '100%',
-                }}
-              >
-                {!!cell?.id && (
-                  <Box
-                    alignItems="center"
-                    display="flex"
-                    flexDirection="column"
-                    gap={1}
-                    justifyContent="center"
-                  >
-                    {!isRestrictedMode && !searching && (
-                      <>
-                        <SelectedPerson orgId={orgId} person={cell} />
-                        <Button
-                          endIcon={<Close />}
-                          onClick={() => updateCellValue(null)}
-                        >
-                          <FormattedMessage id="misc.views.cells.localPerson.clearLabel" />
-                        </Button>
-                      </>
-                    )}
-                  </Box>
-                )}
-                <List
-                  className={styles.searchingList}
-                  sx={{ display: showPeopleInView ? 'block' : 'none' }}
-                >
-                  {showPeopleInView && !!peopleInView.length && (
-                    <List>
-                      <ListSubheader>
-                        <FormattedMessage id="misc.views.cells.localPerson.alreadyInView" />
-                      </ListSubheader>
-                      {peopleInView.map((option) => (
-                        <PersonListItem
-                          key={option.id}
-                          itemProps={{
-                            onClick: () => {
-                              updateCellValue(option);
-                            },
-                          }}
-                          orgId={orgId}
-                          person={option}
-                        />
-                      ))}
-                    </List>
-                  )}
-                  {searching && (
-                    <List {...autoComplete.getListboxProps()}>
-                      <ListSubheader sx={{ position: 'relative' }}>
-                        <FormattedMessage id="misc.views.cells.localPerson.otherPeople" />
-                      </ListSubheader>
-                      {options.map((option, index) => {
-                        const optProps = autoComplete.getOptionProps({
-                          index,
-                          option,
-                        });
-                        return (
-                          <PersonListItem
-                            key={option.id}
-                            itemProps={optProps}
-                            orgId={orgId}
-                            person={option}
-                          />
-                        );
-                      })}
-                    </List>
-                  )}
-                </List>
-              </Box>
-            </Box>
-          )}
-        </Paper>
-      </Popper>
-    </Box>
-  );
-};
-
-const SelectedPerson: FC<{ orgId: number; person: ZetkinPerson }> = ({
-  orgId,
-  person,
-}) => {
-  return (
-    <>
-      <ZUIAvatar orgId={orgId} personId={person.id} size="lg" />
-      <Typography>{`${person.first_name} ${person.last_name}`}</Typography>
-    </>
-  );
-};
-
-const PersonListItem: FC<{
-  itemProps: HTMLAttributes<HTMLLIElement>;
-  orgId: number;
-  person: ZetkinPerson;
-}> = ({ itemProps, orgId, person }) => {
-  return (
-    <ListItem
-      {...itemProps}
-      disablePadding
-      sx={{ paddingBottom: 0.5, paddingTop: 0.5 }}
-    >
-      <Box
-        sx={{
-          cursor: 'pointer',
-          display: 'flex',
-          gap: 1,
-          justifyContent: 'center',
-        }}
-      >
-        <ZUIAvatar orgId={orgId} personId={person.id} size="sm" />
-        <Typography component="span">
-          {`${person.first_name} ${person.last_name}`}
-        </Typography>
-      </Box>
-    </ListItem>
+    <ZUIPersonGridEditCell
+      cell={cell}
+      onGetSuggestedPeople={getPeopleInView}
+      onUpdate={updateCellValue}
+      removePersonLabel="misc.views.cells.localPerson.clearLabel"
+    />
   );
 };
